% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptideLevel_DE.R
\name{plot_3_pep_trends_NOfile}
\alias{plot_3_pep_trends_NOfile}
\title{Plot peptide trends}
\usage{
plot_3_pep_trends_NOfile(mm, prot.info, sorted_norm_m, sorted_prot.info, imp_mm,
  imp_prot.info, prot_to_plot, prot_to_plot_col, gene_name, gene_name_col,
  mylabs)
}
\arguments{
\item{mm}{matrix of raw intensities}

\item{prot.info}{metadata for the intensities in mm}

\item{sorted_norm_m}{normalized intensities, possibly fewer than in mm dew to
filtering out peptides with fewer than one observation per treatment group}

\item{sorted_prot.info}{metadata for the intensities in sorted_norm_m}

\item{imp_mm}{imputed intensities (post normalization)}

\item{imp_prot.info}{metadata for the imputed intensities in imp_mm}

\item{prot_to_plot}{protein ID to plot}

\item{prot_to_plot_col}{protein ID column index}

\item{gene_name}{gene ID to plot}

\item{gene_name_col}{gene ID to plot column index}

\item{mylabs}{sample labels to be plotted on x-axis}
}
\description{
Plot Raw, Normalized and Normalized & Imputed peptide trends for a protein
}
